refactor(clap): update args and build command outside of `main.rs`
This is a work in progress.
