+++ Cargo.toml
@@ -5,7 +5,7 @@ edition = "2021"
 rust-version = "1.75.0"
 
 [dependencies]
-clap = "4.5"
+clap = { version = "4.5", features = ["cargo"] }
 serde = { version = "1", features = [ "derive" ] }
 serde_json = "1"
 erased-serde = "0"
