+++ src/main.rs
@@ -51,6 +51,7 @@ use async_stream::stream;
 use mysql::{params, Pool, prelude::Queryable};
 
 mod errors;
+mod clap;
 
 static mut SERVE_TYPE: Option<ServeType> = None;
 static mut TARGET_DIR: Option<PathBuf> = None;
@@ -420,76 +421,9 @@ fn main() {
 
     log::init(APP_NAME.to_string()).unwrap();
 
-    let args = clap::App::new("wharfix")
-    .arg(clap::Arg::with_name("path")
-        .long("path")
-        .help("Path to directory of static docker image specs")
-        .takes_value(true)
-        .required_unless_one(&["repo", "dbconnfile", "derivationoutput"]))
-    .arg(clap::Arg::with_name("repo")
-        .long("repo")
-        .help("URL to git repository")
-        .takes_value(true)
-        .required_unless_one(&["path", "dbconnfile", "derivationoutput"]))
-    .arg(clap::Arg::with_name("derivationoutput")
-        .long("derivation-output")
-        .help("Output which servable derivations need to produce to be valid")
-        .takes_value(true)
-        .required_unless_one(&["path", "repo", "dbconnfile"]))
-    .arg(clap::Arg::with_name("target")
-        .long("target")
-        .help("Target path in which to checkout repos")
-        .default_value("/tmp/wharfix")
-        .required(false))
-    .arg(clap::Arg::with_name("blobcachedir")
-        .long("blob-cache-dir")
-        .help("Directory in which to store persitent symlinks to docker layer blobs")
-        .takes_value(true)
-        .required(false))
-    .arg(clap::Arg::with_name("substituters")
-        .long("substituters")
-        .help("Comma-separated list of nix substituters to pass directly to nix-build as 'substituters'")
-        .takes_value(true)
-        .required(false))
-    .arg(clap::Arg::with_name("indexfilepath")
-        .long("index-file-path")
-        .help("Path to repository index file")
-        .default_value("default.nix")
-        .takes_value(true)
-        .required(false))
-    .arg(clap::Arg::with_name("indexfileisbuildable")
-        .long("index-file-is-buildable")
-        .help("Set if the provided index-file is a valid nix entrypoint by itself (i.e. don't use internal drv-wrapper)")
-        .takes_value(false)
-        .required(false))
-    .arg(clap::Arg::with_name("sshprivatekey")
-        .long("ssh-private-key")
-        .help("Path to optional ssh private key file")
-        .takes_value(true)
-        .required(false))
-    .arg(clap::Arg::with_name("addnixgcroots")
-        .long("add-nix-gcroots")
-        .help("Whether to add nix gcroots for blobs cached in blob cache dir")
-        .takes_value(false)
-        .required(false)
-        .requires("blobcachedir"))
-    .arg(clap::Arg::with_name("address")
-        .long("address")
-        .help("Listen address to open on <port>")
-        .default_value("0.0.0.0")
-        .required(false))
-    .arg(clap::Arg::with_name("port")
-        .long("port")
-        .help("Listen port to open on <address>")
-        .default_value("8088")
-        .required(true));
+    let matches = clap::build_cli().get_matches();
 
-    #[cfg(feature = "mysql")]
-    let args = args.arg(clap::Arg::with_name("dbconnfile")
-        .long("dbconnfile")
-        .help("Path to file from which to read db connection details")
-        .takes_value(true)
-        .required_unless_one(&["path", "repo", "derivationoutput"]));
+    // TODO: here once lived some parsing definitions.
 
     if let Err(e) = || -> Result<(), MainError> {
